+++ src/main/java/com/sidematch/backend/config/oauth/oauth2user/CustomOAuth2UserService.java
@@ -1,5 +1,6 @@
-package com.sidematch.backend.config.oauth;
+package com.sidematch.backend.config.oauth.oauth2user;
 
+import com.sidematch.backend.config.oauth.OAuth2UserInfoDto;
 import com.sidematch.backend.domain.user.Role;
 import com.sidematch.backend.domain.user.User;
 import com.sidematch.backend.domain.user.service.UserService;
